fix(via-router): use SmallVec instead of StackVec
